backend/requirements.txt
Flask==2.1.1
Flask_Cors==3.0.10
flask_expects_json==1.7.0
jsonschema==4.4.0
PyJWT==2.3.0
pymongo==4.1.0
python-dotenv==0.20.0
